backend/app/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
infobip-api-python-sdk==5.0.1
pydantic==1.10.13
python-dotenv==1.0.0
regex==2022.3.15
requests==2.31.0
typing_extensions==4.8.0
urllib3==2.1.0
